src/main/java/net/binout/codestory2013/scalaskel/Scalaskel.java
package net.binout.codestory2013.scalaskel;

import java.util.*;

public class Scalaskel {

    public Set<ScalaskelResult> change(int nbCents) {
        Set<ScalaskelResult> toReturn = new LinkedHashSet<ScalaskelResult>();

        ScalaskelResult withFoo  = new ScalaskelResult();
        withFoo.setFoo(nbCents);

        toReturn.add(withFoo);

        if (nbCents == 7) {
            ScalaskelResult withBar  = new ScalaskelResult();
            withBar.setBar(1);
            toReturn.add(withBar);
        }

        if (nbCents > 7) {
             int nbBar = (nbCents / 7) -1;
             if (nbBar == 0) nbBar = 1;
             for (int i = 1; i<=nbBar; i++) {
                 Set<ScalaskelResult> barChange = change(nbCents - 7 * i);
                 for (ScalaskelResult result : barChange) {
                     result.addBar(i);
                 }
                 toReturn.addAll(barChange);
             }
        }

        if (nbCents == 11) {
            ScalaskelResult withQix  = new ScalaskelResult();
            withQix.setQix(1);
            toReturn.add(withQix);
        }

        if (nbCents > 11) {
            int nbQix = (nbCents / 11) -1;
            if (nbQix == 0) nbQix = 1;
            for (int i = 1; i<=nbQix; i++) {
                Set<ScalaskelResult> qixChange = change(nbCents - 11 * i);
                for (ScalaskelResult result : qixChange) {
                    result.addQix(i);
                }
                toReturn.addAll(qixChange);
            }
        }

        if (nbCents == 21) {
            ScalaskelResult withBaz  = new ScalaskelResult();
            withBaz.setBaz(1);
            toReturn.add(withBaz);
        }

        if (nbCents > 21) {
            int nbBaz = (nbCents / 21) -1;
            if (nbBaz == 0) nbBaz = 1;
            for (int i = 1; i<=nbBaz; i++) {
                Set<ScalaskelResult> bazChnage = change(nbCents - 21 * i);
                for (ScalaskelResult result : bazChnage) {
                    result.addBaz(i);
                }
                toReturn.addAll(bazChnage);
            }
        }

        return toReturn;
    }
}
